Mapping stats to tables, consensus fix and blast output reduced.
Implemented the collection of statistics regarding the mapping to each target. Now for each target we get statistics about the coverage and the number of nucleotides called in the consensus.

Samtools consensus call got a -a to also print Ns in the beginning and at the end.

The blastn call only needs to print the first result, since we are not using the others anyways. Hence the output was reduced to the first hit